requirements/base.txt
alembic==1.14.1
bcrypt==4.3.0
celery[redis]==5.4.0
email-validator==2.2.0
fastapi==0.115.6
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
redis==5.2.1
sqlalchemy==2.0.36
uvicorn[standard]==0.32.1
